requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.3
anthropic==0.18.1
openai-whisper==20231117
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
aiofiles==23.2.1
pydub==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
